#430: Move ContainerType
